refactor: split wandb settings and metadata
