+++ tools/measurement/measurement_tools/wandb_models.py
@@ -4,248 +4,87 @@
 
 from __future__ import annotations
 
-import hashlib
-import re
 from collections.abc import Mapping
 from enum import StrEnum
-from ipaddress import ip_address
 from pathlib import Path
 from types import MappingProxyType
-from typing import Annotated, Any, Literal, cast
-from urllib.parse import urlsplit
+from typing import Annotated, Literal, cast
 
 from pydantic import (
     BaseModel,
-    ConfigDict,
     Field,
-    FiniteFloat,
     StrictBool,
     StrictInt,
     StrictStr,
     field_validator,
     model_validator,
 )
-from pydantic_settings import BaseSettings, SettingsConfigDict
 
 from measurement_tools.validation import (
     MeasurementStrategy,
     NonNegativeFloat,
     NonNegativeInt,
-    Percentage,
     Probability,
-    RatBenchAttackerEndpointKind,
-    RedactedStrictFrozenModel,
     StrictFrozenModel,
     VisibleIdentifier,
-    VisibleSlurmIdentifier,
+)
+from measurement_tools.wandb_metadata import (
+    BenchmarkMetadata,
+    ConfigMetadata,
+    DetectMetadata,
+    ExecutionMetadata,
+    GitMetadata,
+    ImportedRunMetadata,
+    MatrixMetadata,
+    ModelSourcesMetadata,
+    ReplaceMetadata,
+    RewriteMetadata,
+    RuntimeMetadata,
+    SafeScalar,
+    SlurmJobMetadata,
+    SlurmMetadata,
+    SweepMetadata,
+    WandbConfigPayload,
+    WandbRunMetadata,
+    WandbTag,
+    WorkloadMetadata,
+    WorkloadSourceMetadata,
 )
 from measurement_tools.wandb_metric_schema import AGGREGATED_MEASUREMENT_FIELDS, BENCHMARK_METRIC_NAMES
+from measurement_tools.wandb_policy import DataClass, Exposure, FieldPolicy
+from measurement_tools.wandb_settings import (
+    DEFAULT_WANDB_PROJECT as DEFAULT_WANDB_PROJECT,
+)
+from measurement_tools.wandb_settings import (
+    WANDB_TAG_MAX_LENGTH as WANDB_TAG_MAX_LENGTH,
+)
+from measurement_tools.wandb_settings import (
+    ResolvedWandbConfig as ResolvedWandbConfig,
+)
+from measurement_tools.wandb_settings import (
+    WandbInputs as WandbInputs,
+)
+from measurement_tools.wandb_settings import (
+    WandbMode,
+)
+from measurement_tools.wandb_settings import (
+    generated_wandb_tag as generated_wandb_tag,
+)
+from measurement_tools.wandb_settings import (
+    is_safe_wandb_tag as is_safe_wandb_tag,
+)
+from measurement_tools.wandb_settings import (
+    wandb_tag_value_is_sensitive as wandb_tag_value_is_sensitive,
+)
 
-DEFAULT_WANDB_PROJECT = "nemo-anonymizer-benchmarks"
 PUBLICATION_COMPLETE_KEY = "publication/complete"
 PUBLICATION_SEAL_DIGEST_KEY = "publication/completion_seal_sha256"
-WANDB_TAG_MAX_LENGTH = 64
-_WANDB_TAG_DIGEST_LENGTH = 12
-_SENSITIVE_WANDB_TAG_PARTS = frozenset({"api_key", "credential", "credentials", "password", "secret", "token"})
-_SENSITIVE_WANDB_TAG_VALUE_PREFIXES = ("sk-", "ghp_", "github_pat_", "glpat-", "xoxb-", "xoxp-", "xoxa-")
-
-
-class WandbMode(StrEnum):
-    online = "online"
-    offline = "offline"
-    disabled = "disabled"
 
 
 class WandbPublicationState(StrEnum):
     created = "created"
     resumed = "resumed"
     already_complete = "already_complete"
-
-
-class WandbInputs(BaseSettings):
-    """Optional operator inputs; dedicated measurement variables only."""
-
-    wandb_mode: WandbMode | None = None
-    wandb_project: str | None = None
-    wandb_base_url: str | None = None
-    wandb_entity: str | None = None
-    wandb_group: str | None = None
-    wandb_job_type: str | None = None
-    wandb_run_name: str | None = None
-    wandb_tags: str | None = None
-    wandb_log_tables: bool | None = None
-
-    model_config = SettingsConfigDict(
-        env_prefix="ANONYMIZER_MEASUREMENT_",
-        extra="forbid",
-        env_ignore_empty=True,
-        populate_by_name=True,
-        hide_input_in_errors=True,
-    )
-
-
-class ResolvedWandbConfig(RedactedStrictFrozenModel):
-    """Immutable, fully resolved publisher configuration."""
-
-    wandb_mode: WandbMode = WandbMode.disabled
-    wandb_project: str = DEFAULT_WANDB_PROJECT
-    wandb_base_url: str | None = None
-    wandb_entity: str | None = None
-    wandb_group: str | None = None
-    wandb_job_type: str | None = None
-    wandb_run_name: str | None = None
-    wandb_tags: str = ""
-    wandb_log_tables: bool = False
-
-    @field_validator(
-        "wandb_project",
-        "wandb_base_url",
-        "wandb_entity",
-        "wandb_group",
-        "wandb_job_type",
-        "wandb_run_name",
-    )
-    @classmethod
-    def nonempty_optional_strings(cls, value: str | None) -> str | None:
-        if value is None:
-            return None
-        stripped = value.strip()
-        if not stripped:
-            raise ValueError("W&B string settings cannot be empty")
-        return stripped
-
-    @field_validator("wandb_base_url")
-    @classmethod
-    def validate_base_url(cls, value: str | None) -> str | None:
-        if value is None:
-            return None
-        parsed = urlsplit(value)
-        if (
-            parsed.scheme not in {"http", "https"}
-            or parsed.hostname is None
-            or parsed.username is not None
-            or parsed.password is not None
-            or parsed.query
-            or parsed.fragment
-        ):
-            raise ValueError("W&B base URL must be a credential-free HTTP(S) URL")
-        if parsed.scheme == "http" and not _is_loopback_host(parsed.hostname):
-            raise ValueError("W&B base URL must use HTTPS unless it targets loopback")
-        return value.rstrip("/")
-
-    @field_validator("wandb_tags")
-    @classmethod
-    def validate_tags(cls, value: str) -> str:
-        tags = [tag for tag in (part.strip() for part in value.split(",")) if tag]
-        if any(not is_safe_wandb_tag(tag) for tag in tags):
-            raise ValueError("W&B tags must be safe identifiers between 1 and 64 characters")
-        return value
-
-    @property
-    def enabled(self) -> bool:
-        return self.wandb_mode != WandbMode.disabled
-
-    @property
-    def effective_wandb_project(self) -> str:
-        return self.wandb_project
-
-    @property
-    def effective_wandb_tags(self) -> list[str]:
-        return [tag for tag in (part.strip() for part in self.wandb_tags.split(",")) if tag]
-
-    @classmethod
-    def from_env_and_overrides(
-        cls,
-        *,
-        defaults: Mapping[str, Any] | None = None,
-        **overrides: Any,
-    ) -> ResolvedWandbConfig:
-        source = WandbInputs()
-        values = dict(defaults or {})
-        values.update(source.model_dump(exclude_none=True))
-        values.update({key: value for key, value in overrides.items() if value is not None})
-        return cls.model_validate(values)
-
-    def validated_update(self, **updates: Any) -> ResolvedWandbConfig:
-        values = self.model_dump()
-        values.update(updates)
-        return type(self).model_validate(values)
-
-
-def _is_loopback_host(hostname: str) -> bool:
-    if hostname.lower() == "localhost":
-        return True
-    try:
-        return ip_address(hostname).is_loopback
-    except ValueError:
-        return False
-
-
-def is_safe_wandb_tag(value: str) -> bool:
-    if not value or len(value) > WANDB_TAG_MAX_LENGTH or "://" in value or value.startswith("/"):
-        return False
-    return not wandb_tag_value_is_sensitive(value)
-
-
-def generated_wandb_tag(namespace: str, value: str) -> str | None:
-    candidate = f"{namespace}:{value}"
-    if wandb_tag_value_is_sensitive(candidate) or "://" in candidate or candidate.startswith("/"):
-        return None
-    if len(candidate) <= WANDB_TAG_MAX_LENGTH:
-        return candidate
-    digest = hashlib.sha256(value.encode("utf-8")).hexdigest()[:_WANDB_TAG_DIGEST_LENGTH]
-    prefix_length = WANDB_TAG_MAX_LENGTH - len(namespace) - len(digest) - 2
-    return f"{namespace}:{value[:prefix_length]}-{digest}"
-
-
-def wandb_tag_value_is_sensitive(value: str) -> bool:
-    normalized = value.lower()
-    if normalized.startswith(_SENSITIVE_WANDB_TAG_VALUE_PREFIXES):
-        return True
-    parts = {part for part in re.split(r"[^a-z0-9]+", normalized) if part}
-    if "api" in parts and "key" in parts:
-        return True
-    return bool(parts & _SENSITIVE_WANDB_TAG_PARTS)
-
-
-class DataClass(StrEnum):
-    operational = "operational"
-    pseudonymous = "pseudonymous"
-    sensitive = "sensitive"
-
-
-class Exposure(StrEnum):
-    local_only = "local_only"
-    aggregate = "aggregate"
-    table_opt_in = "table_opt_in"
-    never = "never"
-
-
-class FieldPolicy(BaseModel):
-    data_class: DataClass
-    exposure: Exposure
-
-    model_config = ConfigDict(extra="forbid", frozen=True)
-
-
-SafeScalar = StrictBool | StrictInt | FiniteFloat | StrictStr
-WandbTag = Annotated[StrictStr, Field(min_length=1, max_length=WANDB_TAG_MAX_LENGTH)]
-_NUMERIC_SWEEP_PARAMETER_TAILS = frozenset(
-    {
-        "detect_gliner_threshold",
-        "detect_validation_excerpt_window_chars",
-        "detect_validation_max_entities_per_call",
-        "replace_digest_length",
-        "rewrite_max_repair_iterations",
-    }
-)
-_BOOLEAN_SWEEP_PARAMETER_TAILS = frozenset(
-    {"emit_telemetry", "evaluate", "replace_normalize_label", "rewrite_strict_entity_protection"}
-)
-_ENUM_SWEEP_PARAMETER_VALUES = {
-    "replace_algorithm": frozenset({"md5", "sha1", "sha256"}),
-    "rewrite_risk_tolerance": frozenset({"high", "low", "minimal", "moderate"}),
-}
 
 
 class WandbInitPayload(StrictFrozenModel):
@@ -259,353 +98,6 @@ class WandbInitPayload(StrictFrozenModel):
     job_type: StrictStr
     entity: StrictStr | None = None
     tags: tuple[WandbTag, ...] = ()
-
-
-class BenchmarkMetadata(StrictFrozenModel):
-    metadata_schema_version: NonNegativeInt | None = None
-    suite_schema_version: NonNegativeInt | None = None
-    wandb_sanitizer_version: NonNegativeInt | None = None
-    measurement_schema_version: NonNegativeInt | None = None
-    suite_id: VisibleIdentifier | None = None
-    workload_count: NonNegativeInt | None = None
-    config_count: NonNegativeInt | None = None
-    matrix_entry_count: NonNegativeInt | None = None
-    case_count: NonNegativeInt | None = None
-    case_retries: NonNegativeInt | None = None
-    case_retry_backoff_sec: NonNegativeFloat | None = None
-    suite_file_hash: StrictStr | None = None
-
-
-class BenchmarkIdentityMetadata(StrictFrozenModel):
-    role: Literal["main-baseline", "release-baseline", "candidate"] | None = None
-    kind: Literal["main", "release", "pr", "branch", "experiment"] | None = None
-    suite_version: StrictStr | None = None
-    branch: StrictStr | None = None
-    commit_sha: Annotated[StrictStr, Field(pattern=r"^[0-9a-f]{40,64}$")] | None = None
-    commit_short: Annotated[StrictStr, Field(pattern=r"^[0-9a-f]{7,16}$")] | None = None
-    pr_number: NonNegativeInt | None = None
-    release_version: StrictStr | None = None
-    anonymizer_config_id: VisibleIdentifier | None = None
-    anonymizer_mode: Literal["replace", "rewrite"] | None = None
-
-    @model_validator(mode="after")
-    def validate_branch_identity(self) -> BenchmarkIdentityMetadata:
-        if self.kind in {"pr", "branch"} and self.pr_number is None:
-            raise ValueError("PR or candidate branch benchmark identity requires pr_number")
-        if (
-            self.commit_sha is not None
-            and self.commit_short is not None
-            and not self.commit_sha.startswith(self.commit_short)
-        ):
-            raise ValueError("commit_short must match the prefix of commit_sha")
-        role_kind = {
-            "main-baseline": {"main"},
-            "release-baseline": {"release"},
-            "candidate": {"pr", "branch", "experiment"},
-        }
-        if self.role is not None and self.kind is not None and self.kind not in role_kind[self.role]:
-            raise ValueError(f"benchmark role {self.role!r} is incompatible with kind {self.kind!r}")
-        return self
-
-
-class SlurmJobMetadata(StrictFrozenModel):
-    role: VisibleSlurmIdentifier
-    job_id: VisibleSlurmIdentifier
-
-
-class SlurmMetadata(StrictFrozenModel):
-    job_id: VisibleIdentifier | None = None
-    array_job_id: VisibleIdentifier | None = None
-    array_task_id: VisibleIdentifier | None = None
-    array_task_count: NonNegativeInt | None = None
-    restart_count: NonNegativeInt | None = None
-    ntasks: NonNegativeInt | None = None
-    job_num_nodes: NonNegativeInt | None = None
-    jobs: Annotated[list[SlurmJobMetadata], Field(max_length=64, exclude_if=lambda value: not value)] = Field(
-        default_factory=list
-    )
-
-
-class ExecutionMetadata(StrictFrozenModel):
-    backend: Literal["local", "slurm"] | None = None
-    export: StrictBool | None = None
-    fail_fast: StrictBool | None = None
-    dd_trace: Literal["none", "last_message", "all_messages"] | None = None
-    dd_task_trace: StrictBool | None = None
-    slurm: SlurmMetadata | None = None
-
-
-class RuntimeMetadata(StrictFrozenModel):
-    anonymizer_version: StrictStr | None = None
-    datadesigner_version: StrictStr | None = None
-    wandb_version: StrictStr | None = None
-    python_version: StrictStr | None = None
-    platform_machine: StrictStr | None = None
-    platform_system: StrictStr | None = None
-
-
-class GitMetadata(StrictFrozenModel):
-    commit: StrictStr | None = None
-    branch: StrictStr | None = None
-    dirty: StrictBool | None = None
-
-
-class ModelSourcesMetadata(StrictFrozenModel):
-    has_model_configs: StrictBool | None = None
-    has_model_providers: StrictBool | None = None
-    has_artifact_path: StrictBool | None = None
-
-
-class WorkloadSourceMetadata(StrictFrozenModel):
-    kind: Literal["local_file", "remote_file"] | None = None
-    suffix: StrictStr | None = None
-
-
-class WorkloadMetadata(StrictFrozenModel):
-    id: VisibleIdentifier | None = None
-    text_column: VisibleIdentifier | None = None
-    has_id_column: StrictBool | None = None
-    has_data_summary: StrictBool | None = None
-    row_limit: NonNegativeInt | None = None
-    row_offset: NonNegativeInt | None = None
-    source: WorkloadSourceMetadata | None = None
-
-
-class DetectMetadata(StrictFrozenModel):
-    entity_label_source: Literal["custom", "default"] | None = None
-    entity_label_count: NonNegativeInt | None = None
-    entity_label_set_hash: StrictStr | None = None
-    gliner_threshold: Probability | None = None
-    validation_max_entities_per_call: NonNegativeInt | None = None
-    validation_excerpt_window_chars: NonNegativeInt | None = None
-
-
-class ReplaceMetadata(StrictFrozenModel):
-    strategy: Literal["annotate", "hash", "redact", "substitute"] | None = None
-    normalize_label: StrictBool | None = None
-    algorithm: Literal["md5", "sha1", "sha256"] | None = None
-    digest_length: NonNegativeInt | None = None
-    has_format_template: StrictBool | None = None
-    has_instructions: StrictBool | None = None
-
-
-class RewriteMetadata(StrictFrozenModel):
-    risk_tolerance: Literal["minimal", "low", "moderate", "high"] | None = None
-    max_repair_iterations: NonNegativeInt | None = None
-    strict_entity_protection: StrictBool | None = None
-    has_privacy_goal: StrictBool | None = None
-    has_protect: StrictBool | None = None
-    has_preserve: StrictBool | None = None
-    has_instructions: StrictBool | None = None
-
-
-class ConfigMetadata(StrictFrozenModel):
-    id: VisibleIdentifier | None = None
-    mode: Literal["replace", "rewrite"] | None = None
-    evaluate: StrictBool | None = None
-    emit_telemetry: StrictBool | None = None
-    detect: DetectMetadata | None = None
-    replace: ReplaceMetadata | None = None
-    rewrite: RewriteMetadata | None = None
-
-
-class MatrixMetadata(StrictFrozenModel):
-    workload: VisibleIdentifier | None = None
-    config: VisibleIdentifier | None = None
-    repetitions: NonNegativeInt | None = None
-
-
-class SweepMetadata(StrictFrozenModel):
-    id: VisibleIdentifier
-    arm_id: VisibleIdentifier
-    params: dict[StrictStr, SafeScalar]
-
-    @field_validator("params")
-    @classmethod
-    def validate_params(cls, params: dict[str, SafeScalar]) -> dict[str, SafeScalar]:
-        for key, value in params.items():
-            if not _sweep_parameter_is_safe(key, value):
-                raise ValueError(f"unsupported W&B sweep parameter: {key}")
-        return params
-
-    @classmethod
-    def from_arm(cls, *, sweep_id: str, arm_id: str, params: dict[str, Any]) -> SweepMetadata:
-        safe_params = {
-            key: value
-            for key, value in params.items()
-            if isinstance(value, bool | int | float | str) and _sweep_parameter_is_safe(key, value)
-        }
-        return cls(id=sweep_id, arm_id=arm_id, params=safe_params)
-
-
-class ImportedRunMetadata(StrictFrozenModel):
-    completion_seal_schema_version: NonNegativeInt
-    completion_seal_sha256: StrictStr = Field(pattern=r"^[0-9a-f]{64}$")
-    producer_repository: VisibleIdentifier
-    producer_commit: StrictStr = Field(pattern=r"^[0-9a-f]{40,64}$")
-    phase: VisibleIdentifier
-    case_id: VisibleIdentifier
-
-
-class WandbRunMetadata(StrictFrozenModel):
-    """Validated benchmark metadata that is safe to project into W&B config."""
-
-    run_kind: Literal["native_suite", "sweep_arm", "imported_case"] = "native_suite"
-    benchmark: BenchmarkMetadata | None = None
-    execution: ExecutionMetadata | None = None
-    runtime: RuntimeMetadata | None = None
-    git: GitMetadata | None = None
-    model_sources: ModelSourcesMetadata | None = None
-    benchmark_identity: BenchmarkIdentityMetadata | None = None
-    workloads: tuple[WorkloadMetadata, ...] = ()
-    configs: tuple[ConfigMetadata, ...] = ()
-    matrix: tuple[MatrixMetadata, ...] = ()
-    sweep: SweepMetadata | None = None
-    imported: ImportedRunMetadata | None = None
-
-    @field_validator("workloads", "configs", "matrix", mode="before")
-    @classmethod
-    def freeze_sequences(cls, value: Any) -> tuple[Any, ...]:
-        if not isinstance(value, list | tuple):
-            raise ValueError("W&B metadata sequences must be lists or tuples")
-        return tuple(value)
-
-    @model_validator(mode="after")
-    def validate_run_kind(self) -> WandbRunMetadata:
-        if (self.sweep is not None) != (self.run_kind == "sweep_arm"):
-            raise ValueError("sweep metadata and sweep_arm run kind must be set together")
-        if (self.imported is not None) != (self.run_kind == "imported_case"):
-            raise ValueError("import metadata and imported_case run kind must be set together")
-        if self.benchmark is None or self.benchmark.suite_id is None:
-            raise ValueError("W&B run metadata requires a benchmark suite identity")
-        if self.run_kind == "imported_case":
-            config_ids = [config.id for config in self.configs if config.id is not None]
-            if len(config_ids) != 1:
-                raise ValueError("imported W&B metadata requires exactly one config identity")
-        return self
-
-    def comparer_values(self) -> dict[str, SafeScalar]:
-        values: dict[str, SafeScalar] = {}
-        benchmark = self.benchmark
-        if benchmark is not None:
-            _set_scalar(values, "benchmark_suite_id", benchmark.suite_id)
-            _set_scalar(values, "benchmark_case_count", benchmark.case_count)
-        _set_scalar(values, "benchmark_workload_ids", _compact_values(item.id for item in self.workloads))
-        _set_scalar(
-            values,
-            "benchmark_workload_row_limits",
-            _compact_values(item.row_limit for item in self.workloads),
-        )
-        _set_scalar(
-            values,
-            "benchmark_workload_source_kinds",
-            _compact_values(item.source.kind if item.source else None for item in self.workloads),
-        )
-        _set_scalar(
-            values,
-            "benchmark_workload_source_suffixes",
-            _compact_values(item.source.suffix if item.source else None for item in self.workloads),
-        )
-        _set_scalar(values, "benchmark_config_ids", _compact_values(item.id for item in self.configs))
-        _set_scalar(values, "benchmark_modes", _compact_values(item.mode for item in self.configs))
-        _set_scalar(values, "benchmark_strategies", _compact_values(_config_strategy(item) for item in self.configs))
-        _set_scalar(
-            values,
-            "benchmark_gliner_thresholds",
-            _compact_values(item.detect.gliner_threshold if item.detect else None for item in self.configs),
-        )
-        _set_scalar(
-            values,
-            "benchmark_entity_label_counts",
-            _compact_values(item.detect.entity_label_count if item.detect else None for item in self.configs),
-        )
-        _set_scalar(
-            values,
-            "benchmark_risk_tolerances",
-            _compact_values(item.rewrite.risk_tolerance if item.rewrite else None for item in self.configs),
-        )
-        identity = self.benchmark_identity
-        if identity is not None:
-            _set_scalar(values, "benchmark_role", identity.role)
-            _set_scalar(values, "benchmark_kind", identity.kind)
-            _set_scalar(values, "suite_version", identity.suite_version)
-            _set_scalar(values, "branch", identity.branch)
-            _set_scalar(values, "commit_sha", identity.commit_sha)
-            _set_scalar(values, "commit_short", identity.commit_short)
-            _set_scalar(values, "pr_number", identity.pr_number)
-            _set_scalar(values, "release_version", identity.release_version)
-            _set_scalar(values, "anonymizer_config_id", identity.anonymizer_config_id)
-            _set_scalar(values, "anonymizer_mode", identity.anonymizer_mode)
-        return values
-
-
-class WandbConfigPayload(WandbRunMetadata):
-    suite_id: VisibleIdentifier
-    wandb_mode: WandbMode
-    wandb_log_tables: StrictBool
-    benchmark_suite_id: VisibleIdentifier | None = None
-    benchmark_case_count: NonNegativeInt | None = None
-    benchmark_workload_ids: SafeScalar | None = None
-    benchmark_workload_row_limits: SafeScalar | None = None
-    benchmark_workload_source_kinds: SafeScalar | None = None
-    benchmark_workload_source_suffixes: SafeScalar | None = None
-    benchmark_config_ids: SafeScalar | None = None
-    benchmark_modes: SafeScalar | None = None
-    benchmark_strategies: SafeScalar | None = None
-    benchmark_gliner_thresholds: SafeScalar | None = None
-    benchmark_entity_label_counts: SafeScalar | None = None
-    benchmark_risk_tolerances: SafeScalar | None = None
-    benchmark_role: Literal["main-baseline", "release-baseline", "candidate"] | None = None
-    benchmark_kind: Literal["main", "release", "pr", "branch", "experiment"] | None = None
-    suite_version: StrictStr | None = None
-    branch: StrictStr | None = None
-    commit_sha: Annotated[StrictStr, Field(pattern=r"^[0-9a-f]{40,64}$")] | None = None
-    commit_short: Annotated[StrictStr, Field(pattern=r"^[0-9a-f]{7,16}$")] | None = None
-    pr_number: NonNegativeInt | None = None
-    release_version: StrictStr | None = None
-    anonymizer_config_id: VisibleIdentifier | None = None
-    anonymizer_mode: Literal["replace", "rewrite"] | None = None
-    native_suite_id: VisibleIdentifier | None = None
-    sweep_id: VisibleIdentifier | None = None
-    sweep_arm_id: VisibleIdentifier | None = None
-    imported_config_id: VisibleIdentifier | None = None
-    sweep_params: dict[StrictStr, SafeScalar] = Field(default_factory=dict)
-
-    def sdk_values(self) -> dict[str, Any]:
-        values = self.model_dump(mode="json", exclude_none=True)
-        values.update({f"sweep_param_{key}": value for key, value in self.sweep_params.items()})
-        return values
-
-    @classmethod
-    def from_run_metadata(
-        cls,
-        settings: ResolvedWandbConfig,
-        *,
-        suite_id: str,
-        metadata: WandbRunMetadata,
-    ) -> WandbConfigPayload:
-        values: dict[str, Any] = {
-            "suite_id": suite_id,
-            "wandb_mode": settings.wandb_mode,
-            "wandb_log_tables": settings.wandb_log_tables,
-        }
-        if metadata.benchmark is None or metadata.benchmark.suite_id != suite_id:
-            raise ValueError("W&B metadata suite identity does not match publication suite")
-        values.update({name: getattr(metadata, name) for name in WandbRunMetadata.model_fields})
-        values.update(metadata.comparer_values())
-        if metadata.run_kind == "native_suite":
-            values["native_suite_id"] = metadata.benchmark.suite_id
-        if metadata.sweep is not None:
-            values.update(
-                {
-                    "sweep_id": metadata.sweep.id,
-                    "sweep_arm_id": metadata.sweep.arm_id,
-                    "sweep_params": metadata.sweep.params,
-                }
-            )
-        if metadata.run_kind == "imported_case":
-            values["imported_config_id"] = metadata.configs[0].id
-        return cls.model_validate(values, strict=True)
 
 
 class WandbHistoryPayload(StrictFrozenModel):
@@ -795,26 +287,6 @@ class TraceCoverageTableRow(StrictFrozenModel):
     unsupported_column_count: NonNegativeInt
 
 
-class RatBenchReidentificationTableRow(_MetricTableRow):
-    record_type: Literal["rat_bench_reidentification"]
-    rows_processed: NonNegativeInt
-    rows_failed: NonNegativeInt
-    reidentified_rows: NonNegativeInt
-    direct_reidentified_rows: NonNegativeInt
-    correctmatch_reidentified_rows: NonNegativeInt
-    reid_error_rows: NonNegativeInt
-    reidentification_rate_pct: Percentage
-    coverage_pct: Percentage
-    correct_guess_count: NonNegativeInt
-    incorrect_guess_count: NonNegativeInt
-    total_guess_count: NonNegativeInt
-    mean_reid_score: NonNegativeFloat
-    reid_threshold: NonNegativeFloat
-    missing_output_rows: NonNegativeInt | None = None
-    attacker_model: StrictStr | None = None
-    attacker_endpoint_kind: RatBenchAttackerEndpointKind | None = None
-
-
 WandbTableRow = Annotated[
     RunTableRow
     | StageTableRow
@@ -822,8 +294,7 @@ WandbTableRow = Annotated[
     | EvaluationTableRow
     | NddWorkflowTableRow
     | ModelWorkflowTableRow
-    | TraceCoverageTableRow
-    | RatBenchReidentificationTableRow,
+    | TraceCoverageTableRow,
     Field(discriminator="record_type"),
 ]
 WANDB_TABLE_ROW_MODELS: Mapping[str, type[BaseModel]] = MappingProxyType(
@@ -835,7 +306,6 @@ WANDB_TABLE_ROW_MODELS: Mapping[str, type[BaseModel]] = MappingProxyType(
         "ndd_workflow": NddWorkflowTableRow,
         "model_workflow": ModelWorkflowTableRow,
         "dd_trace_coverage": TraceCoverageTableRow,
-        "rat_bench_reidentification": RatBenchReidentificationTableRow,
     }
 )
 _MEASUREMENT_RECORD_TYPES = frozenset(WANDB_TABLE_ROW_MODELS)
@@ -898,47 +368,6 @@ class WandbPublishResult(StrictFrozenModel):
     record_count: StrictInt = 0
 
 
-def _set_scalar(mapping: dict[str, SafeScalar], key: str, value: Any) -> None:
-    if isinstance(value, bool | int | float | str):
-        mapping[key] = value
-
-
-def _compact_values(values: Any) -> SafeScalar | None:
-    compacted: list[SafeScalar] = []
-    seen: set[tuple[str, SafeScalar]] = set()
-    for value in values:
-        if not isinstance(value, bool | int | float | str):
-            continue
-        marker = (type(value).__name__, value)
-        if marker in seen:
-            continue
-        compacted.append(value)
-        seen.add(marker)
-    if len(compacted) == 1:
-        return compacted[0]
-    if len(compacted) > 1:
-        return ",".join(str(value) for value in compacted)
-    return None
-
-
-def _config_strategy(config: ConfigMetadata) -> str | None:
-    if config.replace is not None:
-        return config.replace.strategy
-    return "rewrite" if config.rewrite is not None else None
-
-
-def _sweep_parameter_is_safe(key: str, value: SafeScalar) -> bool:
-    if not key.startswith("configs_") or not key.isascii() or any(not (char.isalnum() or char == "_") for char in key):
-        return False
-    if any(key.endswith(f"_{tail}") for tail in _NUMERIC_SWEEP_PARAMETER_TAILS):
-        if isinstance(value, bool) or not isinstance(value, int | float) or value < 0:
-            return False
-        return not key.endswith("_detect_gliner_threshold") or value <= 1
-    if any(key.endswith(f"_{tail}") for tail in _BOOLEAN_SWEEP_PARAMETER_TAILS):
-        return isinstance(value, bool)
-    return any(key.endswith(f"_{tail}") and value in allowed for tail, allowed in _ENUM_SWEEP_PARAMETER_VALUES.items())
-
-
 def _aggregate_policies(
     *names: str,
     data_class: DataClass = DataClass.operational,
@@ -973,19 +402,6 @@ OUTBOUND_FIELD_POLICIES: dict[type[BaseModel], dict[str, FieldPolicy]] = {
         "case_retries",
         "case_retry_backoff_sec",
         "suite_file_hash",
-    ),
-    BenchmarkIdentityMetadata: _aggregate_policies(
-        "role",
-        "kind",
-        "suite_version",
-        "branch",
-        "commit_sha",
-        "commit_short",
-        "pr_number",
-        "release_version",
-        "anonymizer_config_id",
-        "anonymizer_mode",
-        data_class=DataClass.pseudonymous,
     ),
     SlurmJobMetadata: _aggregate_policies("role", "job_id", data_class=DataClass.pseudonymous),
     SlurmMetadata: _aggregate_policies(
@@ -1060,7 +476,6 @@ OUTBOUND_FIELD_POLICIES: dict[type[BaseModel], dict[str, FieldPolicy]] = {
         "runtime",
         "git",
         "model_sources",
-        "benchmark_identity",
         "workloads",
         "configs",
         "matrix",
@@ -1084,16 +499,6 @@ OUTBOUND_FIELD_POLICIES: dict[type[BaseModel], dict[str, FieldPolicy]] = {
         "benchmark_gliner_thresholds",
         "benchmark_entity_label_counts",
         "benchmark_risk_tolerances",
-        "benchmark_role",
-        "benchmark_kind",
-        "suite_version",
-        "branch",
-        "commit_sha",
-        "commit_short",
-        "pr_number",
-        "release_version",
-        "anonymizer_config_id",
-        "anonymizer_mode",
         "native_suite_id",
         "sweep_id",
         "sweep_arm_id",
@@ -1156,27 +561,6 @@ OUTBOUND_FIELD_POLICIES: dict[type[BaseModel], dict[str, FieldPolicy]] = {
         "native_trace_column_count",
         "private_trace_column_count",
         "unsupported_column_count",
-        exposure=Exposure.table_opt_in,
-    ),
-    RatBenchReidentificationTableRow: _aggregate_policies(
-        "record_type",
-        "rows_processed",
-        "rows_failed",
-        "reidentified_rows",
-        "direct_reidentified_rows",
-        "correctmatch_reidentified_rows",
-        "reid_error_rows",
-        "reidentification_rate_pct",
-        "coverage_pct",
-        "correct_guess_count",
-        "incorrect_guess_count",
-        "total_guess_count",
-        "mean_reid_score",
-        "reid_threshold",
-        "missing_output_rows",
-        "attacker_model",
-        "attacker_endpoint_kind",
-        data_class=DataClass.operational,
         exposure=Exposure.table_opt_in,
     ),
     WandbTablePayload: {
